Find/requirements.txt
telethon==1.23.0
PySocks
pymysql